market reports refactored and extracted. created version 2 of formMap constant
